% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-group.R
\name{age_category}
\alias{age_category}
\title{Create age categories}
\usage{
age_category(dt = NULL, interval = NULL)
}
\arguments{
\item{dt}{Dataset}

\item{interval}{Age interval}
}
\description{
Create age categories either by interval or specified age categories.
}
\examples{
\dontrun{
dd
}
}
