% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{debug_opt}
\alias{debug_opt}
\title{Debugging options}
\usage{
debug_opt(opt = c("fun", "nrow", "row", "aggregate", "geo"), val = NULL)
}
\arguments{
\item{opt}{Debug options}

\item{val}{Value options}
}
\description{
A wrapper for debugging options.
}
\details{
Options available:
\itemize{
\item "fun"       : Activating options \code{orgdata.debug}
\item "nrow"      : For \code{orgdata.debug.nrow}. Default value is 20 if \code{val} argument is missing
\item "row"       : For \code{orgdata.debug.row}. Default value is 1:50 if \code{val} argument is missing
\item "aggregate" : Activating \code{orgdata.debug.aggregate}
\item "geo"       : Activating \code{orgdata.debug.geo}
}
}
\examples{
\dontrun{
debug_opt("geo")
debug_opt("nrow", 30)
}
}
