% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-columns.R
\name{do_addcols}
\alias{do_addcols}
\title{Add Columns}
\usage{
do_addcols(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{cols}{Old and new columns. See output from \code{get_addcols()}}
}
\description{
Add columns to the dataset
}
