% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate_recode}
\alias{do_aggregate_recode}
\title{Recode Aggregated Variables}
\usage{
do_aggregate_recode(dt)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}
}
\description{
Recode aggregated variables to represent all values either as \code{0}
integer variables and \code{Tot} for string variables.
}
