% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate_recode_standard}
\alias{do_aggregate_recode_standard}
\title{Recode Standard Aggregated Variables}
\usage{
do_aggregate_recode_standard(dt)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}
}
\description{
Recode standard aggregated variables to represent total category either as \code{0}
or \code{10} for integer variables and \code{Tot} for string variables. Value \code{10} representing
total is only used for \code{LANDB} since it already has \code{0} as one of it's existing
value. Other columns that need to be recorded to represent total category must be defined
in \strong{Recode} form with FILGRUPPE as \strong{AGGREGATE}.

Standard columns that is implemented by this function is:

\code{UTDANN}, \code{SIVILSTAND}, \code{LANDB} and \code{LANDSSB}
}
\seealso{
Other aggregate functions: 
\code{\link{do_aggregate}()},
\code{\link{get_aggregate}()}
}
\concept{aggregate functions}
