% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-columns.R
\name{do_colname}
\alias{do_colname}
\title{Rename Columns}
\usage{
do_colname(dt = NULL, cols = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{cols}{Old and new columns. See output from \code{get_colname()}}
}
\description{
Rename standard columns in the dataset
}
