% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-column.R
\name{do_column_standard}
\alias{do_column_standard}
\title{Rename Standard Columns}
\usage{
do_column_standard(dt = NULL, spec = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specification data as list. See output from
\code{get_column_standard}}
}
\description{
Renaming standard columns. To see what are the standard columns
run \code{orgdata:::is_standard_cols()}.
}
