% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-column.R
\name{do_column_standard_rename}
\alias{do_column_standard_rename}
\title{Rename Standard Columns}
\usage{
do_column_standard_rename(file = NULL, spec = NULL)
}
\arguments{
\item{file}{Input data}

\item{spec}{Specifications data as a data.frame}
}
\description{
Renaming standard columns as in \code{getOption("orgdata.columns")}.
}
