% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-row.R
\name{do_delete_row}
\alias{do_delete_row}
\title{Delete Selected Rows}
\usage{
do_delete_row(dt = NULL, spec = NULL, con = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Delete selected rows in the dataset. The rows to be deleted can be specified directly
in the Access registration database using minus symbol \code{-}
}
\seealso{
Other delete-row functions: 
\code{\link{get_delete_row_spec}()}
}
\concept{delete-row functions}
