% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-args.R
\name{do_extra_args}
\alias{do_extra_args}
\title{Execute Extra Arguments}
\usage{
do_extra_args(dt = NULL, args = NULL)
}
\arguments{
\item{dt}{Dataset}

\item{args}{Extra arguments as specified in details section below.}
}
\description{
This is based on the input in \code{EXTRA} column from Access
registration database. The arguments that are valid here can be expanded
whenever needed. See details section for valid arguments to be used.
All argument names are written in \code{CamelCase} style.
}
\details{
Currently, these arguments can be used:
\itemize{
\item \code{DeleteNaRow} : Delete any row consisting only NA in all columns
}
}
\seealso{
Other extra arguments: 
\code{\link{get_extra_args}()}
}
\concept{extra arguments}
