% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo-recode.R
\name{do_geo_recode}
\alias{do_geo_recode}
\title{Recode Geographical Codes}
\usage{
do_geo_recode(
  dt = NULL,
  code = NULL,
  type = c("grunnkrets", "fylke", "kommune", "bydel"),
  year = NULL,
  con = NULL,
  geo = getOption("orgdata.debug.geo"),
  base = getOption("orgdata.recode.base")
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{code}{Code dataset of old and new codes in a \code{data.table} format.}

\item{type}{The geographical granularity for recoding The dataset is the
output after running \code{get_geo_recode()} function.}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then current year will be used.}

\item{con}{Connection to database}

\item{geo}{Keep old geographical code if TRUE. Default is FALSE.}

\item{base}{Logical. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}
}
\description{
Recode geographical codes to the current year. Codes is based on
\code{\link[norgeo:track_change]{norgeo::track_change()}} function. For a split geogaphical codes from
previous year, the first code of the current year code in chronological order
will be selected to recode.
}
\examples{
\dontrun{
code <- get_geo_recode(con = geo$dbconn, type = "grunnkrets")
dt <- make_file("BEFOLKNING", aggregate = FALSE)
DT <- do_geo_recode(dt, code)
}
}
\seealso{
Other geo recode functions: 
\code{\link{do_recode_without_aggregate}()},
\code{\link{get_geo_recode}()}
}
\concept{geo recode functions}
