% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manheader.R
\name{do_manheader}
\alias{do_manheader}
\title{Rename Colname Manually}
\usage{
do_manheader(dt = NULL, manspec = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{manspec}{Specification from \code{MANHEADER} column in \code{tbl_Innlesing}}
}
\description{
Rename the columns manually based on column index.
}
