% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-column.R
\name{do_mutate}
\alias{do_mutate}
\title{Create Column}
\usage{
do_mutate(dt = NULL, spec = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specifications data as a data.frame}
}
\description{
Create new column with the value when specified between a
symbols less then \code{<} and more than \code{>} in the standard columns. For
example when specified under \code{KJONN} as \verb{<2>}, a new column \code{KJONN} is
created with value 2.

This function is applicable for columns \verb{"GEO", "AAR", "KJONN", "ALDER", "UTDANN", "LANDSSB"}, all \code{"TABS"} and all \code{"VALS"}.
}
\seealso{
Other mutate-column functions: 
\code{\link{get_mutate_value}()}
}
\concept{mutate-column functions}
