% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{do_norgeo}
\alias{do_norgeo}
\title{Create Georgraphical Codes}
\usage{
do_norgeo(year = NULL, write = FALSE, append = FALSE)
}
\arguments{
\item{year}{Year for geographical codes}

\item{write}{Write table \code{tblGeo} to the \code{orgdata.geo} database. It will overwrite
the \code{tblGeo} if it already exists}

\item{append}{Append the data to an existing \code{tblGeo} in the \code{orgdata.geo}}
}
\description{
Create a database for geographical codes to aggregate
data accordingly. Implementation of this function is base on \code{\link[norgeo:cast_geo]{norgeo::cast_geo()}}
function in \href{https://helseprofil.github.io/norgeo/}{norgeo} package.
}
