% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-post.R
\name{do_recode_post}
\alias{do_recode_post}
\title{Post Recode Columns}
\usage{
do_recode_post(dt = NULL, spec = NULL, con = NULL, control = FALSE)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specification from \code{tbl_Filgruppe}}

\item{con}{Connection to database}

\item{control}{Logical value. \code{TRUE} if the file has been controlled for
possible errors}
}
\description{
Recode selected columns after cleaning and processing the
filegroup based on the specification for \code{FILGRUPPE} in codebook with type
\code{PS}. For example some columns in a filegroup doesn't have values for a
specific year. This column should not have a value of 0. Instead a special
symbol \code{..} is given, representing that the column is missing because the
value is not available to calculate the value.
}
\seealso{
Other recode functions: 
\code{\link{do_recode_aggregate}()},
\code{\link{do_recode_regexp}()},
\code{\link{do_recode}()},
\code{\link{get_codebok_aggregate}()},
\code{\link{get_codebok_regexp}()},
\code{\link{get_codebok}()}
}
\concept{recode functions}
