% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-regexp.R
\name{do_recode_regexp}
\alias{do_recode_regexp}
\title{Recode Variables with Regular Expression}
\usage{
do_recode_regexp(dt = NULL, spec = NULL, con = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Recode variables with regular expression based on the
specification in \code{tbl_KodeBok} with TYPE \code{RE}. The specification in the
codebook can be specific, common or general. \emph{Specific} is when both
\code{LESID} and \code{FILGRUPPE} are specified to create a unique reference to be
able to recode the variables. \emph{Common} is when only \code{FILGRUPPE} is
specified while \code{LESID} left empty and \emph{general} is when \code{FILGRUPPE}
is specified with \code{ALLE} while \code{LESID} left empty.
}
\seealso{
Other recode functions: 
\code{\link{do_recode_aggregate}()},
\code{\link{do_recode}()},
\code{\link{get_codebok_aggregate}()},
\code{\link{get_codebok_regexp}()},
\code{\link{get_codebok}()}
}
\concept{recode functions}
