% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo-recode.R
\name{do_recode_without_aggregate}
\alias{do_recode_without_aggregate}
\title{Recode Geo Code Without Aggregate}
\usage{
do_recode_without_aggregate(
  dt = NULL,
  source = c("grunnkrets", "fylke", "kommune", "bydel"),
  year = NULL,
  check = getOption("orgdata.debug.aggregate"),
  base = getOption("orgdata.recode.base")
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{source}{What geographical granularity codes that is available in the
source data. This will be used for merging with the output from
\code{geo_levels()}}

\item{year}{Which year the georaphical code is valid for. If not specified,
then it will be base on the year in source data ie. column \code{AAR}}

\item{check}{If TRUE then output will keep variables for geographical levels
without aggregating it. This is useful to check for geographical codes that
are missing. Else use \code{options(orgdata.aggregate = FALSE)}}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}
}
\description{
Recode geo code without aggregating the data.
The input in argument \code{source} must be a lower
granularity level than the \code{level} input.
}
\examples{
\dontrun{
# Source data with enumeration area codes ie. grunnkrets
dt <- make_file("BEFOLKNING", aggregate = FALSE)
}
}
\seealso{
Other geo recode functions: 
\code{\link{do_geo_recode}()},
\code{\link{get_geo_recode}()}
}
\concept{geo recode functions}
