% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{do_reshape}
\alias{do_reshape}
\title{Reshape from Wide to Long}
\usage{
do_reshape(dt = NULL, respec = NULL)
}
\arguments{
\item{dt}{Dataset to be reshaped}

\item{respec}{Reshape specification with \code{id} and \code{measure} variables. This
is the output from \code{get_reshape_id_val()}}
}
\description{
Reshape the dataset from wide format to long format.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape_rename_col}()},
\code{\link{do_reshape_wide}()},
\code{\link{get_reshape_id_val}()},
\code{\link{get_reshape_wide_spec}()}
}
\concept{reshape functions}
