% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-wide.R
\name{do_reshape_long}
\alias{do_reshape_long}
\title{Special Need Reshape to Long}
\usage{
do_reshape_long(dt, resval, rescol, widecols)
}
\arguments{
\item{dt}{An output dataset from \code{do_reshape_wide()}}

\item{resval}{Column value to be reshaped from}

\item{rescol}{Column(s) dimension where the value derived from}

\item{widecols}{Column(s) names created for wide reshape in the dataset}
}
\description{
This function is only applicable to reshape data that was
reshaped to wide via Access specification in \code{RESHAPE} columns.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape_rename_col}()},
\code{\link{do_reshape_wide}()},
\code{\link{do_reshape}()},
\code{\link{get_reshape_id_val}()},
\code{\link{get_reshape_wide_spec}()}
}
\concept{reshape functions}
