% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{do_reshape_rename_col}
\alias{do_reshape_rename_col}
\title{Rename Reshaped Column}
\usage{
do_reshape_rename_col(dt = NULL, spec = NULL)
}
\arguments{
\item{dt}{Dataset that has been reshaped}

\item{spec}{Specification in column \code{RESHAPE_KOL} as in \code{tbl_Innlesing}}
}
\description{
Reshaping variables in the dataset from wide to long will
produce ID and value columns. This function will rename these columns when
specified under \code{RESHAPE_KOL} in the Access registration database.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape}()},
\code{\link{get_reshape_id_val}()}
}
\concept{reshape functions}
