% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-wide.R
\name{do_reshape_wide}
\alias{do_reshape_wide}
\title{Reshape from Long to Wide}
\usage{
do_reshape_wide(dt = NULL, respec = NULL)
}
\arguments{
\item{dt}{Dataset to be reshaped}

\item{respec}{Reshape specification with \code{id} and \code{value} variables. This
is the output from \code{get_reshape_wide_spec()}}
}
\description{
Reshape the dataset from long format to wide format.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape_rename_col}()},
\code{\link{do_reshape}()},
\code{\link{get_reshape_id_val}()},
\code{\link{get_reshape_wide_spec}()}
}
\concept{reshape functions}
