% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source-file.R
\name{do_source_file}
\alias{do_source_file}
\title{Source R or Stata files}
\usage{
do_source_file(dt, extra)
}
\arguments{
\item{dt}{Dataset}

\item{extra}{Arguments from EXTRA column in Access}
}
\description{
Sourcing R or Stata files via \code{EXTRA} column. All files should
be in \emph{source} repo in Helseprofil Github site and saved in their
respective folders. \strong{This function is not implemented yet.}
}
