% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-column.R
\name{do_split}
\alias{do_split}
\title{Split Column}
\usage{
do_split(dt = NULL, split = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{split}{Split specification as a list. Should be equivalent
to the output of \code{\link[=get_split]{get_split()}} function}
}
\description{
Split the columns as specified in the registration database.
}
