% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{do_year}
\alias{do_year}
\title{Add Year to Data}
\usage{
do_year(dt = NULL, year = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{year}{Output from \code{\link[=get_year]{get_year()}}}
}
\description{
Add column for year if it doesn't exist in the rawdata
}
