% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{emoji}
\alias{emoji}
\title{Emoji}
\usage{
emoji(
  x = c("mark", "thumb", "write", "smile", "sad", "santa", "search", "folder", "book")
)
}
\arguments{
\item{x}{Emoji to choose ie. thumb, smile etc}
}
\description{
Change emoji in the output messages for fun 😄
}
\examples{
emoji("smile")
}
