% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{find_column_input}
\alias{find_column_input}
\title{Columns with Single Input}
\usage{
find_column_input(
  spec = NULL,
  col = NULL,
  type = c("character", "double", "integer")
)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{col}{Column name in the database table}

\item{type}{Type of object as output ie. double, integer or character.
Default is character.}
}
\description{
Get the value of a column when it's only one input is allowed.
}
