% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column-rename.R
\name{find_column_standard}
\alias{find_column_standard}
\title{Rename standard columns}
\usage{
find_column_standard(spec = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}
}
\value{
A list with \code{old} and \code{new} columnnames
}
\description{
Rename columns in the rawdata to the standard names
as in \code{getOptions("orgdata.columns")}.
}
