% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-data-file.R
\name{find_data}
\alias{find_data}
\title{Find Rawdata}
\usage{
find_data(file = NULL, ...)
}
\arguments{
\item{file}{Input file}

\item{...}{Other options relevant to the file type}
}
\description{
Find rawdata and load it into memory for further process.
It requires the class attribute of the file is set prior to
using \code{\link[=find_data]{find_data()}} method. Otherwise, use \code{\link[=read_file]{read_file()}} function
instead unless you have specific reason not to.
}
\examples{
\dontrun{
file <- "path/to/my-data.csv"
class(file) <- "csv"
dt <- find_data(file)
}
}
