% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit-null.R
\name{find_implicit_null}
\alias{find_implicit_null}
\title{Dataset for Implicit Null}
\usage{
find_implicit_null(imp, year, colstr, level)
}
\arguments{
\item{imp}{A list of implicit null data derived from \link{get_implicit_col} function}

\item{year}{Selected year from the dataset}

\item{colstr}{Column structure or Class type}

\item{level}{Geographical granularity for aggregating data to}
}
\description{
Create a dataset with implicit null for every selected columns
}
\seealso{
Other implicit-null functions: 
\code{\link{do_implicit_null}()},
\code{\link{find_implicit_col}()},
\code{\link{get_implicit_col}()},
\code{\link{get_implicit_per_year}()},
\code{\link{get_implicit_ref}()}
}
\concept{implicit-null functions}
