% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesing-year.R
\name{find_year}
\alias{find_year}
\title{Year for The Data}
\usage{
find_year(id = NULL, con = NULL)
}
\arguments{
\item{id}{\code{FILID} from table \emph{tbl_Orgfile}}

\item{con}{Connection to database}
}
\value{
An integer
}
\description{
Find the year value in column \code{DEFAAR} from table \emph{tbl_Orgfile}
when column for year isn't available in the raw data. This is
indicated with \code{$Y} in column \code{AAR} in table \emph{tbl_Innlesing}.
}
