% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{geo_map}
\alias{geo_map}
\title{Granularity of Georgraphical Codes}
\usage{
geo_map(year = NULL, write = FALSE, append = FALSE, table = "tblGeo")
}
\arguments{
\item{year}{Year for the valid geographical codes}

\item{write}{Write table to the \code{orgdata.geo} database. It will overwrite
the table if it already exists}

\item{append}{Append the data to an existing table in the \code{orgdata.geo}}

\item{table}{Table name to be created in the database. Default is \code{tblGeo}}
}
\description{
Create a database granularity of geographical codes to aggregate
data accordingly. Implementation of this function is base on \code{\link[norgeo:cast_geo]{norgeo::cast_geo()}}
function in \href{https://helseprofil.github.io/norgeo/}{norgeo} package.
}
\examples{
\dontrun{
geo_map(2020, write = TRUE)
geo_map(2021, append = TRUE)
}
}
\seealso{
Other geo codes functions: 
\code{\link{geo_recode}()},
\code{\link{get_grunnkrets_dummy}()}
}
\concept{geo codes functions}
