% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{geo_merge}
\alias{geo_merge}
\title{Merge Other Geo Level Manually}
\usage{
geo_merge(
  id.table = NULL,
  id.file = NULL,
  geo.col = NULL,
  geo.level = NULL,
  geo.name = NULL,
  file = NULL,
  year = NULL,
  write = FALSE,
  table.name = "tblGeo",
  ...
)
}
\arguments{
\item{id.table}{ID columname to merge to that is found in the database eg.
\code{kommune}}

\item{id.file}{ID columname from the file to merge from. This depends on the
columnames in the files. If \code{id.table} is \code{kommune}, then \code{id.file} must be
the columname representing geo codes that is equivalent to \code{kommune} codes.
Both \code{id.table} and \code{id.file} will be used for merging and these codes must
be unique.}

\item{geo.col}{Columname where the new geo codes are eg. if the new geo codes
is on levekaar, then \code{geo.col} is the columname where codes for levekaar
can be found.}

\item{geo.level}{Geographical level of the merged file will be representing
eg. "levekaar". This will be the value in column \code{level} in the \code{tblGeo} in
the database.
the value in column \code{name} in the \code{tblGeo}in the database.}

\item{geo.name}{Column containing names of the geographical units in the merged
file. This will be the value in column \code{name}in the \code{tblGeo} in the database.}

\item{file}{Complete path of filename to merge from}

\item{year}{Year the code is valid for. If not sepecified \code{orgdata.year} is
used.}

\item{write}{}

\item{table.name}{Name of the table for geo recode in geocodes database. This
can be found with \code{getOptions("orgdata.geo")}. The default is \code{tblGeo}.}

\item{...}{Other possible arguments}
}
\description{
Geo codes other than those downloaded from SSB API can be merged
to the main geo table ie. \code{tblGeo} in the geocodes database. The file must
consist of id column to merge into ie. \code{id.file} and the geo codes to add
to ie. \code{geo.col}.
}
\examples{
\dontrun{
dt <- geo_merge(id.table = "grunnkrets",
                id.file = "id",
                geo.col = "col2",
                geo.level = "levekaar",
                geo.name = "col3",
                file = "C:/path/to/file.csv",
                year = 2022)
}
}
