% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{geo_recode}
\alias{geo_recode}
\title{Geographical Codes}
\usage{
geo_recode(
  type = c("grunnkrets", "bydel", "kommune", "fylke"),
  from = NULL,
  to = NULL,
  write = FALSE,
  append = FALSE
)
}
\arguments{
\item{type}{Type of regional granularity ie. enumeration area (grunnkrets)}

\item{from}{Starting year for the range period. Current year is the default if left empty}

\item{to}{End of year for the range period. Current year is the default if left empty}

\item{write}{Write table to the \code{orgdata.geo} database. It will overwrite
the table if it already exists}

\item{append}{Append the data to an existing table in the \code{orgdata.geo}}
}
\description{
Create a table of current year geographical codes against previous
years geogprahical codes. This is used to recode the previous years codes to the
current codes. Implementation of this function is base on \code{\link[norgeo:track_change]{norgeo::track_change()}}
function.
}
\examples{
\dontrun{
geo_recode(type = "grunnkrets", from = 2018, to = 2021, write = TRUE)
geo_recode(type = "grunnkrets", from = 2018, to = 2021, append = TRUE)
}
}
\seealso{
Other geo codes functions: 
\code{\link{geo_map}()},
\code{\link{get_grunnkrets_dummy}()}
}
\concept{geo codes functions}
