% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-aggregate.R
\name{get_aggregere}
\alias{get_aggregere}
\title{Aggregate Data}
\usage{
get_aggregere(group = NULL, con = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}
}
\description{
Get the specification on how the data will be aggregated to
different geographical levels ie. county, manucipality, town etc.
}
