% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{get_codebok}
\alias{get_codebok}
\title{Codebook}
\usage{
get_codebok(spec = NULL, con = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Get the codebook for recoding variables based on the
unique \code{LESID} number.
}
