% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-aggregate.R
\name{get_codebok_aggregate}
\alias{get_codebok_aggregate}
\title{Codebook for Aggregate Columns}
\usage{
get_codebok_aggregate(spec = NULL, con = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Get the codebook for recoding of aggregated variables based on
the \code{FILGRUPPE}. Specification group \code{ALLE} will be used when \code{FILGRUPPE}
is not specified.
}
\seealso{
Other recode functions: 
\code{\link{do_recode_aggregate}()},
\code{\link{do_recode_regexp}()},
\code{\link{do_recode}()},
\code{\link{get_codebok_regexp}()},
\code{\link{get_codebok}()}
}
\concept{recode functions}
