% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-regexp.R
\name{get_codebok_regexp}
\alias{get_codebok_regexp}
\title{Codebook with Regular Expression}
\usage{
get_codebok_regexp(spec = NULL, con = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Get the codebook with regular expression to recode variables
based on the \code{FILGRUPPE} and \code{LESID} number. Specification group \code{ALLE}
will be used when neither \code{FILGRUPPE} nor \code{LESID} is specified.
}
\seealso{
Other recode functions: 
\code{\link{do_recode_aggregate}()},
\code{\link{do_recode_regexp}()},
\code{\link{do_recode}()},
\code{\link{get_codebok_aggregate}()},
\code{\link{get_codebok}()}
}
\concept{recode functions}
