% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-columns.R
\name{get_colname}
\alias{get_colname}
\title{Get Renamed Columns}
\usage{
get_colname(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification from \code{tbl_Filgruppe}}
}
\value{
A list consist of two variables ie. \code{old} and \code{new}
indicating the old and new columnames
}
\description{
Get the old and new coloumnames to be renamed in the dataset.
}
