% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{get_column_input}
\alias{get_column_input}
\title{Columns with Single Input}
\usage{
get_column_input(df, col, type = c("double", "integer", "character"))
}
\arguments{
\item{df}{Data.frame extracted form registration database}

\item{col}{Column name in the database table}

\item{type}{Type of object as output ie. double, integer or character}
}
\description{
Get the value of a column when it's only one input is allowed.
}
