% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_column_multi_args}
\alias{get_column_multi_args}
\alias{get_input_multi_args}
\title{Columns with Multiple Argument}
\usage{
get_column_multi_args(col = NULL, sep = c(",", ":", ";"))

get_input_multi_args(arg = NULL, y = NULL)
}
\arguments{
\item{col}{Column name in the table}

\item{sep}{Symbols that seperate these arguments eg. "," or ":"}

\item{arg}{Name of the arguments}

\item{y}{Argument(s) as a vector}
}
\description{
Get all arguments in the columns that have multiple arguments with \code{get_column_multi_args()}.
The output will be a vector that can be used in \code{get_input_multi_args()} to get the value of the input.
}
\examples{
\dontrun{
args <- get_column_multi_args(spec$INNLESARG)
val <- get_input_multi_arg("header", args)
}
}
