% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-column.R
\name{get_column_standard}
\alias{get_column_standard}
\title{Get Standard Columns}
\usage{
get_column_standard(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The name of filegroup as specified in \emph{filgruppe}}

\item{con}{Connection to database}

\item{spec}{Specification of the standard columns in \code{tbl_Innlesing}}
}
\value{
A list with \code{old} and \code{new} columnnames
}
\description{
Standard columns names in rawdata will be checked against
the standard names in options as in \code{getOption("orgdata.columns")}.
Nevertheless column \code{GEO} is a special case when geo codes are derived
from a combination of two columns.
}
