% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-row.R
\name{get_delete_row_spec}
\alias{get_delete_row_spec}
\title{Get Rows for Deletion}
\usage{
get_delete_row_spec(spec = NULL, con = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Get the raws to be deleted as specified with minus \code{-} symbol
in the codebook based on the \code{FILGRUPPE} and \code{LESID} number.
Specification group \code{ALLE} will be used when neither \code{FILGRUPPE} nor \code{LESID} is specified.
}
\seealso{
Other delete-raw functions: 
\code{\link{do_delete_row}()}
}
\concept{delete-raw functions}
