% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra-args.R
\name{get_extra_args}
\alias{get_extra_args}
\title{Get Extra Arguments}
\usage{
get_extra_args(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specifications data as a data.frame}
}
\value{
A list with the names and value of arguments
}
\description{
This is based on the input in \code{EXTRA} column from Access registration database.
The arguments that are valid here can be expanded whenever needed. All argument names are written
in \code{CamelCase} style. Use comma to seperate multiple arguments.
}
\seealso{
Other extra arguments: 
\code{\link{do_extra_args_group}()},
\code{\link{do_extra_args}()},
\code{\link{find_age_category}()},
\code{\link{get_extra_args_group}()}
}
\concept{extra arguments}
