% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{get_geo_dummy}
\alias{get_geo_dummy}
\title{Create Dummy Enumeration Area Codes}
\usage{
get_geo_dummy(dt, from, to, type, fix = TRUE)
}
\arguments{
\item{dt}{Downloaded data with norgeo::track_change()}

\item{from}{Starting year for the range period. Current year is the default if left empty}

\item{to}{End of year for the range period. Current year is the default if left empty}

\item{type}{Type of regional granularity ie. enumeration area (grunnkrets)}

\item{fix}{Default is TRUE. Use external codes to fix geo
changes manually. The codes is sourced from
\href{https://github.com/helseprofil/backend/tree/main/norgeo}{config} files
depending on the granularity levels.}
}
\description{
Some of the downloaded enumeration area codes from SSB lack
codes for missing ie. \code{xxxx9999}. This function create these codes that
are needed for recoding older codes to current enumeration area codes.
}
\seealso{
Other geo codes functions: 
\code{\link{geo_map}()},
\code{\link{geo_recode}()}
}
\concept{geo codes functions}
