% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo-recode.R
\name{get_geo_recode}
\alias{get_geo_recode}
\title{Get Previous and Current Geo Codes}
\usage{
get_geo_recode(
  con = NULL,
  type = c("grunnkrets", "fylke", "kommune", "bydel"),
  year = NULL
)
}
\arguments{
\item{con}{Connection to database}

\item{type}{The geographical granularity to recode}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then current year will be used.}
}
\value{
A dataset with columns \code{GEO} and \code{to} representing the GEO codes that
will be recoded to a new code ie. \code{to}.
}
\description{
Get the geographical codes registered in \code{geo-database} which
consist of old and new codes that are applicable to the respective year.
}
\seealso{
Other geo recode functions: 
\code{\link{do_geo_recode}()},
\code{\link{do_recode_without_aggregate}()}
}
\concept{geo recode functions}
