% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit-null.R
\name{get_implicit_col}
\alias{get_implicit_col}
\title{Get Implicit Null Categories}
\usage{
get_implicit_col(dt, years, cols, refs)
}
\arguments{
\item{dt}{Dataset consisting all years}

\item{years}{All years in the dataset}

\item{cols}{Columns to be controlled for implicit null}

\item{refs}{Reference values for the selected columns}
}
\description{
Get categories with implicit values for each selected columns
}
\seealso{
Other implicit-null functions: 
\code{\link{do_implicit_null}()},
\code{\link{find_implicit_col}()},
\code{\link{find_implicit_null}()},
\code{\link{get_implicit_per_year}()},
\code{\link{get_implicit_ref}()}
}
\concept{implicit-null functions}
