% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit-null.R
\name{get_implicit_per_year}
\alias{get_implicit_per_year}
\title{Dataset for Implicit Null by Year}
\usage{
get_implicit_per_year(imp, refs, years, colstr)
}
\arguments{
\item{imp}{A list of implicit null data derived from \link{get_implicit_col} function}

\item{refs}{Reference values for the selected columns}

\item{years}{All years in the dataset}

\item{colstr}{Column structure or Class type}
}
\description{
Create a dataset with implicit null for each available
year in the dataset
}
\seealso{
Other implicit-null functions: 
\code{\link{do_implicit_null}()},
\code{\link{find_implicit_col}()},
\code{\link{find_implicit_null}()},
\code{\link{get_implicit_col}()},
\code{\link{get_implicit_ref}()}
}
\concept{implicit-null functions}
