% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implicit-null.R
\name{get_implicit_ref}
\alias{get_implicit_ref}
\title{Get Reference Values}
\usage{
get_implicit_ref(dt, cols)
}
\arguments{
\item{dt}{Dataset consisting all years}

\item{cols}{Columns to be controlled for implicit null}
}
\description{
Get the valid values for columns that will be controlled for
any possible implicit null.
}
\seealso{
Other implicit-null functions: 
\code{\link{do_implicit_null}()},
\code{\link{find_implicit_col}()},
\code{\link{find_implicit_null}()},
\code{\link{get_implicit_col}()},
\code{\link{get_implicit_per_year}()}
}
\concept{implicit-null functions}
