% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesing.R
\name{get_innlesarg}
\alias{get_innlesarg}
\title{Get File Specifications}
\usage{
get_innlesarg(arg, col, ...)
}
\arguments{
\item{arg}{Name of the arguments}

\item{col}{Column name in the table}

\item{...}{Other arguments eg. sep = ":"}
}
\value{
A character
}
\description{
This function will access all specifications on how
the data will be clean, restructured and aggregated. These specifications are registered in
the following register database tables:
\enumerate{
\item{tbl_Filgruppe}
\item{tbl_Orgfile}
\item{tbl_Innlesing}
}
}
\examples{
\dontrun{
input <- get_innlesarg("ark", spec$INNLESARG)
}
}
