% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-functions.R
\name{get_manheader}
\alias{get_manheader}
\title{Get MANHEADER}
\usage{
get_manheader(df = NULL)
}
\arguments{
\item{df}{Specifications data as a data.frame}
}
\value{
A list containing \code{$index} to refer to the column index and
\code{$col} for the new name of the selected column index.
}
\description{
Get the inputs for column MANHEADER in \code{tbl_Innlesing}.
The input tells that we want to manually \strong{rename} the column for
various reasons such as column name in the rawdata is too long or
it uses unstandard naming style. An input like this:

\code{3,6=AGE,EDUCATION}

means we want to rename column \code{3} to \code{AGE} and column \code{6}
to \code{EDUCATION}.
}
