% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-column.R
\name{get_mutate_value}
\alias{get_mutate_value}
\title{Value for New Column}
\usage{
get_mutate_value(col)
}
\arguments{
\item{col}{The selected column}
}
\description{
Get the value for the new column if the column
fullfilled the condition as a new column.
}
\seealso{
Other mutate-column functions: 
\code{\link{do_mutate}()}
}
\concept{mutate-column functions}
