% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_reshape_id_val}
\alias{get_reshape_id_val}
\title{Reshape Id and Measure}
\usage{
get_reshape_id_val(dt = NULL, group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{dt}{Dataset to be reshaped}

\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification from \code{tbl_Filgruppe}}
}
\description{
Get the id and measure variables for reshaping the dataset. For
detail please read \code{data.table::melt.data.table} to understand \code{id} and
\code{mearsure} variables.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape_rename_col}()},
\code{\link{do_reshape}()}
}
\concept{reshape functions}
