% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-wide.R
\name{get_reshape_wide_spec}
\alias{get_reshape_wide_spec}
\title{Reshape Wide Specification}
\usage{
get_reshape_wide_spec(dt = NULL, group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{dt}{Dataset to be reshaped}

\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification from \code{tbl_Filgruppe}}
}
\description{
Get the formula and value variables for reshaping a long dataset
to wide. For detail please read \code{data.table::dcast.data.table} to
understand \code{formula} and \code{value} variables.
}
\seealso{
Other reshape functions: 
\code{\link{do_reshape_long}()},
\code{\link{do_reshape_rename_col}()},
\code{\link{do_reshape_wide}()},
\code{\link{do_reshape}()},
\code{\link{get_reshape_id_val}()}
}
\concept{reshape functions}
