% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{get_spec}
\alias{get_spec}
\title{Get File Specification}
\usage{
get_spec(filgruppe = NULL, con = NULL)
}
\arguments{
\item{filgruppe}{The \emph{filgruppe} of files category}

\item{con}{Connection to database}
}
\value{
A data.frame
}
\description{
This function will access all specifications on where the orginal data is and how
the data will be restructured and aggregated. The specifications are registered in
the following register database:
\enumerate{
\item{tbl_Koble}
\item{tbl_Orgfile}
\item{tbl_Innlesing}
}
}
