% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-column.R
\name{get_split}
\alias{get_split}
\title{Get Split Column}
\usage{
get_split(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The name of filegroup as specified in \emph{filgruppe}}

\item{con}{Connection to database}

\item{spec}{Specification from \code{tbl_Filgruppe}}
}
\value{
A list consist of two variables ie. \code{from} and \code{to}
indicating which column to split and what the new column names will be
}
\description{
Get the column to be splitted and the column name
for the new splitted columns.
If you already have the specification from \code{\link[=find_spec]{find_spec()}}, then the arguments
for \code{group} and \code{con} must be \code{NULL}.
}
