% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesing.R
\name{get_year_from_file}
\alias{get_year_from_file}
\alias{get_year}
\title{Get year}
\usage{
get_year_from_file(id = NULL, con = NULL)

get_year(col = NULL)
}
\arguments{
\item{id}{Is \code{FILID} from table \emph{tbl_Orgfile}}

\item{con}{Connection to database}
}
\value{
An integer
}
\description{
Get the year value in column \code{DEFAAR} from table \emph{tbl_Orgfile}
when column for year isn't available in the raw data. This is
indicated with \code{$Y} in the table \emph{ tbl_Innlesing }.
}
