% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{input_type}
\alias{input_type}
\title{Type of object input}
\usage{
input_type(value, type = c("double", "integer", "character"))
}
\arguments{
\item{value}{Input value}
}
\description{
Convert value to selected type ie. checked with \link{typeof}
}
\keyword{internal}
