% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{is_active}
\alias{is_active}
\title{Use Original Columnames}
\usage{
is_active(active = getOption("orgdata.active"), .env = parent.frame())
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is either to use original columnames or use the standard columnames.
It's deprecated because output MUST use standard columnames.
}
\keyword{internal}
