% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{is_col_separate}
\alias{is_col_separate}
\title{Multiple inputs style}
\usage{
is_col_separate(input, int = NULL)
}
\arguments{
\item{input}{The product from \code{find_column_input()}}
}
\description{
Multiple inputs separated by \verb{,} as in ADDKOL and MANHEADER
}
\keyword{internal}
