% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column-rename.R
\name{is_column_standard}
\alias{is_column_standard}
\title{Rename standard columns}
\usage{
is_column_standard(df)
}
\arguments{
\item{df}{Data.frame extracted form registration database}
}
\value{
A list with \code{old} and \code{new} columnnames
}
\description{
Rename columns in the rawdata to the standard names
as in \code{getOptions("orgdata.columns")}.
}
\keyword{internal}
