% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-db-connect.R
\name{is_conn_db}
\alias{is_conn_db}
\title{Connect to Database}
\usage{
is_conn_db(dbname = NULL, db = c("kh", "geo"), .test = FALSE)
}
\arguments{
\item{dbname}{Database filename with complete path}

\item{db}{Database file \code{kh} (Kommunehelse) and \code{geo} (Geo code)}

\item{.test}{Use for testing only}
}
\description{
Use R6 object to connect to database
}
\keyword{internal}
