% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year.R
\name{is_defaar}
\alias{is_defaar}
\title{Year for The Data}
\usage{
is_defaar(id = NULL, con = NULL)
}
\arguments{
\item{id}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{con}{Connection to database}
}
\value{
An integer
}
\description{
Find the year value in column \code{DEFAAR} from table \emph{tbl_Orgfile}
when column for year isn't available in the raw data. This is
indicated with \code{$Y} in column \code{AAR} in table \emph{tbl_Innlesing}.
}
\keyword{internal}
