% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{is_input_type}
\alias{is_input_type}
\title{Type of object input}
\usage{
is_input_type(value, type = c("character", "double", "integer"))
}
\arguments{
\item{value}{Input value to be converted}
}
\description{
Convert value to selected type ie. checked with \link{typeof}
}
\keyword{internal}
