% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-org.R
\name{is_process_org}
\alias{is_process_org}
\title{Process Raw Data}
\usage{
is_process_org(file, filespec, fgspec, con)
}
\arguments{
\item{file}{File rawdata}

\item{filespec}{Specification for a file from \code{tbl_Innlesing}}

\item{fgspec}{Specification for a file group}

\item{con}{Connection to the database}
}
\value{
A dataset with \code{data.table} format
}
\description{
Implement the specifications to the raw data.
}
\keyword{internal}
