% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-spec.R
\name{is_query}
\alias{is_query}
\title{Create the SQL query}
\usage{
is_query(
  file = NULL,
  value = NULL,
  external = FALSE,
  char = NULL,
  char2 = NULL,
  opposite = FALSE
)
}
\description{
Create the SQL query to get the data as specified
}
\keyword{internal}
