% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6-db-connect.R
\name{KHelse}
\alias{KHelse}
\title{R6 Class Connecting to Database}
\description{
Connect to registration database to get all necessary information
on data source and cleaning process.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbname}}{Database filename.}

\item{\code{dbconn}}{Database connection.}

\item{\code{tblname}}{Table name to be created in the database.}

\item{\code{tblvalue}}{Data to be inserted in the table \code{tblname}.
Data must be in a \code{data.frame} or \code{data.table} format.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KHelse$new()}}
\item \href{#method-db_connect}{\code{KHelse$db_connect()}}
\item \href{#method-db_write}{\code{KHelse$db_write()}}
\item \href{#method-db_close}{\code{KHelse$db_close()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Start connecting to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$new(dbname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbname}}{Database filename.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_connect"></a>}}
\if{latex}{\out{\hypertarget{method-db_connect}{}}}
\subsection{Method \code{db_connect()}}{
Reconnect to the database when \code{db_close} was used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_write"></a>}}
\if{latex}{\out{\hypertarget{method-db_write}{}}}
\subsection{Method \code{db_write()}}{
Write table to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_write(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name to be created in the database.}

\item{\code{value}}{The data to be inserted in the table.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db_close"></a>}}
\if{latex}{\out{\hypertarget{method-db_close}{}}}
\subsection{Method \code{db_close()}}{
Close connection to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_close()}\if{html}{\out{</div>}}
}

}
}
