% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\docType{data}
\name{log}
\alias{log}
\title{Environment to store log info}
\format{
An object of class \code{environment} of length 0.
}
\usage{
log
}
\description{
Environment to store log info
}
\keyword{datasets}
