% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-group.R
\name{make_filegroup}
\alias{make_filegroup}
\alias{lag_filgruppe}
\title{Make Multiple Filegroups}
\usage{
make_filegroup(...)

lag_filgruppe(...)
}
\arguments{
\item{...}{Filegroup(s)}
}
\description{
Implement the specification in registration database on several
filegroups at once. This function will be only used when all selected
filegroups have been controlled for any errors. The selected filegroups
will be save as a \code{.csv}.
}
\examples{
\dontrun{
make_filegroup(NEET, TRANGBODD, DODE)
}
}
\seealso{
Other filegroups functions: 
\code{\link{make_file}()}
}
\concept{filegroups functions}
