% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file.R
\name{make_raw}
\alias{make_raw}
\title{Aggregate Rawdata Asis}
\usage{
make_raw(file, geo, val)
}
\arguments{
\item{file}{Input file}

\item{geo}{Which columname has geo codes. Only use in inactive mode.
Check status with \code{getOption("orgdata.active")}}

\item{val}{Which colunames has the value for quantity. Only use in an inactive mode.
Check status with \code{getOption("orgdata.active")}}
}
\description{
Aggregate rawdata as it is ie. keeping the columnames unchanged.
}
\keyword{internal}
