% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdata-package.R
\docType{package}
\name{orgdata-package}
\alias{orgdata}
\alias{orgdata-package}
\title{orgdata: Aggregating Original Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Cleaning, restructuring and aggregating original data usually with enumerical area codes to any geographical codes.
}
\details{
Aggregating Original Data (raw data) into different geographical
levels such as county, manucipality or town.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://helseprofil.github.io/orgdata}
  \item Report bugs at \url{https://github.com/helseprofil/orgdata/issues}
}

}
\author{
\strong{Maintainer}: Vegard Lysne \email{vegard.lysne@helsedir.no}

Authors:
\itemize{
  \item Yusman Kamaleri \email{ybka@fhi.no} (\href{https://orcid.org/0000-0001-5014-3665}{ORCID})
}

Other contributors:
\itemize{
  \item FHI KHelse group (https://www.fhi.no/hn/folkehelse) [contributor, reviewer]
}

}
\keyword{internal}
