% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdata-package.R
\docType{package}
\name{orgdata}
\alias{orgdata}
\alias{orgdata-package}
\title{Aggregating Original Data into Geographical Levels}
\description{
Aggregating Origianl Data (raw data) into different geographical
levels such as county, manucipality or town
}
\keyword{internal}
