% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{orgdata_logo}
\alias{orgdata_logo}
\title{Orgdata logo}
\usage{
orgdata_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Use utf-8. Default: cli::is_utf8_output()}
}
\value{
Orgdata logo
}
\description{
Display orgdata logo
}
\examples{
\dontrun{
if(interactive()){
 print.orgdata_logo()
 }
}
}
\seealso{
\code{\link[cli]{is_utf8_output}}, \code{\link[cli]{ansi-styles}}
}
