% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\docType{data}
\name{orgEnv}
\alias{orgEnv}
\title{Status workflow}
\format{
An object of class \code{environment} of length 1.
}
\usage{
orgEnv
}
\description{
Status workflow to display message. Default is 1
}
\keyword{datasets}
