% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-file.R
\name{read_file}
\alias{read_file}
\title{Read Data File}
\usage{
read_file(file = NULL, ...)
}
\arguments{
\item{file}{Input file}

\item{...}{Other options relevant to the file type}
}
\description{
Read rawdata. It uses the \code{\link[=find_data]{find_data()}} generic method.
For a \code{ .csv } file, \code{\link[data.table:fread]{data.table::fread()}} is used and all the arguments
for \code{fread} function can be used. For a \code{ xlsx } or \code{ .xls } file
\code{\link[readxl:read_excel]{readxl::read_excel()}} function is used and all its arguments.
}
\examples{
\dontrun{
rdata <- read_file("/file/path/mydata.xlsx", sheet = "S3", range = cell_rows(1:4))
rdata <- read_file("/file/path/mydata.csv", sep = ",", header = FALSE)
}
}
