% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{read_log}
\alias{read_log}
\title{Read Log File}
\usage{
read_log(name = NULL, koblid = NULL)
}
\arguments{
\item{name}{Log filename without file extention}

\item{koblid}{Connecting ID or \code{KOBLID}}
}
\description{
Read the log file from orgdata default path. To check the
default path use \code{orgdata:::is_orgdata_path()}.
}
\examples{
\dontrun{
# To list codes with xxxx9999 of koblid 67
read_log("code99", 67)
}
}
