% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-org.R
\name{read_org}
\alias{read_org}
\alias{lesorg}
\title{Read and Run the Specifications}
\usage{
read_org(group = NULL, koblid = NULL)

lesorg(group = NULL, koblid = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}
}
\description{
Read the specifications in the register database and implement them
to the selected group of files (\emph{filgruppe}). All files under
the selected group will be considered unless the \code{KOBLID} with
argument \code{koblid} is specified. Specifying \code{koblid} is useful for testing.

The function \code{\link[=lesorg]{lesorg()}} is an
alias to \code{\link[=read_org]{read_org()}}.
}
\examples{
\dontrun{
read_org("BEFOLKNING")
read_org("BEFOLKNING", koblid = 19)
read_org("BEFOLKNING", koblid = c(15, 50))
}
}
