% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-raw.R
\name{read_raw}
\alias{read_raw}
\alias{lesraw}
\title{Read and Implement the Specifications}
\usage{
read_raw(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  geo = NULL,
  val = NULL,
  ...
)

lesraw(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  geo = NULL,
  val = NULL,
  ...
)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{aggregate}{Aggregate data according to the specification in registration database.
Default is FALSE. Use \code{options(orgdata.aggregate = TRUE)} to change globally.}

\item{save}{Save as \code{.csv} by activating \code{save_file()}. Default is \code{FALSE}}

\item{year}{Which year the georaphical code is valid for. If not specified, then
it will be base on the year in source data ie. column \code{AAR}}

\item{geo}{Which columname has geo codes. Only use in inactive mode.
Check status with \code{getOption("orgdata.active")}}

\item{val}{Which colunames has the value for quantity. Only use in an inactive mode.
Check status with \code{getOption("orgdata.active")}}

\item{...}{Additional parameters}
}
\description{
Read the specifications in the register database and implement them
to the raw data of the selected group of files (\emph{filgruppe}). All files under
the selected group will be affected unless the \code{KOBLID} with
argument \code{koblid} is specified. Specifying \code{koblid} is useful
especially for testing purposes.

The function \code{\link[=lesraw]{lesraw()}} is an
alias to \code{\link[=read_raw]{read_raw()}}.
}
\examples{
\dontrun{
read_raw("BEFOLKNING")
read_raw("BEFOLKNING", koblid = 19)
read_raw("BEFOLKNING", koblid = c(15, 50))
}
}
