% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-data.R
\name{see_file}
\alias{see_file}
\alias{se_fil}
\title{See Column Categories}
\usage{
see_file(dt = NULL, ...)

se_fil(dt = NULL, ...)
}
\arguments{
\item{dt}{Dataset of type \code{data.frame} or \code{data.table}}

\item{...}{Columnames to be displayed}
}
\description{
See what categories available in the data.
}
\examples{
\dontrun{
see_file(DT, KJONN, UTDANN, LANDSSB)
}
}
