% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\name{see_raw}
\alias{see_raw}
\title{See Raw Data in the Database}
\usage{
see_raw(name = NULL, filid = NULL, action = c("read", "delete"))
}
\arguments{
\item{name}{The filegroup name}

\item{filid}{File ID for the raw data}

\item{action}{To read or delete the data in the database. Default is \code{read}.}
}
\description{
See the raw data that are saved in the raw database when the
column \emph{KONTROLLERT} is marked.
}
\examples{
\dontrun{
dt <- see_raw("LESEFERD", filid = 134)
}
}
