% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{update_khpackage}
\alias{update_khpackage}
\alias{upgrade_khpackage}
\title{Update khpackage}
\usage{
update_khpackage(
  package = c("orgdata", "orgcube", "qualcontrol"),
  branch = "main",
  force = FALSE
)

upgrade_khpackage(
  package = c("orgdata", "orgcube", "qualcontrol"),
  branch = "main",
  force = FALSE
)
}
\arguments{
\item{package}{Which khpackage to update}

\item{branch}{The branch in Github to install from, default = "main"}

\item{force}{Use the latest version(s) of all dependencies. Default is FALSE}
}
\description{
Update package directly with Github repo. Default is \code{main} branch.
}
\examples{
\dontrun{
update_khpackage(package = "orgdata") #default
update_khpackage(package = "orgdata", branch = "dev") #to upgrade with dev branch
}
}
