% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{website}
\alias{website}
\title{Show Website}
\usage{
website(url = NULL)
}
\arguments{
\item{url}{The URL}
}
\description{
Display website of \code{orgdata} package ie. \href{https://helseprofil.github.io/orgdata/}{https://helseprofil.github.io/orgdata/}
}
