% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Aggregate Data}
\usage{
do_aggregate(
  dt = NULL,
  source = c("G", "F", "K", "B"),
  level = c("G", "F", "K", "B")
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{source}{Geo code available in the source data for merging}

\item{level}{Geographical levels for aggregating data}
}
\description{
Aggregate data according to the specification in \code{tbl_Filgruppe}
}
