% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Aggregate Data}
\usage{
do_aggregate(
  dt = NULL,
  source = c("grunnkrets", "fylke", "kommune", "bydel"),
  level = c("grunnkrets", "fylke", "kommune", "bydel"),
  year = NULL,
  check = FALSE
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{source}{What geographical granularity code that is available in the source data.
This will be used for merging with the output from \code{do_norgeo()}}

\item{level}{Geographical granularity for aggregating data.}

\item{year}{Which year the georaphical code is valid for. If not specified, then
it will be base on the year in source data ie. column \code{AAR}}

\item{check}{If TRUE then output will not be aggregated. This is useful to check
for geographical codes that are missing.}
}
\description{
Aggregate data according to the specification in \code{tbl_Filgruppe}.
}
\examples{
\dontrun{
# To aggregate source data with enumeration area codes ie. grunnkrets, to
# manucipaltiy ie. kommune
dt <- read_org("BEFOLKNING")
DT <- do_aggregate(dt, source = "grunnkrets", level = "kommune")
}
}
