% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Aggregate Data}
\usage{
do_aggregate(
  dt = NULL,
  source = c("grunnkrets", "fylke", "kommune", "bydel"),
  level = c("grunnkrets", "fylke", "kommune", "bydel")
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{source}{What geographical granularity code that is available in the source data.
This will be used for merging with the output from \code{do_norgeo()}}

\item{level}{Geographical granularity for aggregating data.}
}
\description{
Aggregate data according to the specification in \code{tbl_Filgruppe}.
}
\examples{
\dontrun{
  # To aggregate source data with enumeration area codes ie. grunnkrets, to
  # manucipaltiy ie. kommune
  DT <- do_aggregate(dt, source = "grunnkrets", level = "kommune")
}
}
