% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{do_aggregate}
\alias{do_aggregate}
\title{Aggregate Data}
\usage{
do_aggregate(
  dt = NULL,
  source = c("grunnkrets", "fylke", "kommune", "bydel"),
  level = getOption("orgdata.geo.levels"),
  year = NULL,
  aggregate.col = NULL,
  geoDT = NULL,
  check = NULL,
  base = NULL,
  control = FALSE,
  wide = NULL
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{source}{What geographical granularity codes that is available in the
source data. This will be used for merging with the geo codebook generated
from \code{geo_map()} ie. from \code{tblGeo} in geo database}

\item{level}{Geographical granularity for aggregating data. See \code{getOption("orgdata.geo.levels")}}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then global option for \code{orgdata.year} will be used.}

\item{aggregate.col}{Other columns to aggregate other than the standard ie.
\code{UTDANN}, \code{LANDSSB}, \code{LANDBAK} and \code{INNVKAT}}

\item{geoDT}{Geo codes to aggregate dataset with}

\item{check}{If TRUE then output will keep variables for geographical levels
without aggregating it. This is useful to check for geographical codes that
are missing. Else use \code{options(orgdata.aggregate = FALSE)}}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}

\item{control}{Logical value. \code{TRUE} if the file has been controlled for
possible errors}

\item{wide}{Column(s) from reshape wide}
}
\description{
Aggregate data according to the specification in \code{tbl_Filgruppe}
in \code{AGGREGERE} column. The input in argument \code{source} must be a lower
granularity level than the \code{level} input.
}
\examples{
\dontrun{
# To aggregate source data with enumeration area codes ie. grunnkrets, to
# manucipaltiy ie. kommune
dt <- make_file("BEFOLKNING")
DT <- do_aggregate(dt, source = "grunnkrets", level = "kommune")
}
}
\seealso{
Other aggregate functions: 
\code{\link{get_aggregate}()}
}
\concept{aggregate functions}
