% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo-recode.R
\name{do_geo_recode}
\alias{do_geo_recode}
\title{Recode Geographical Codes}
\usage{
do_geo_recode(dt = NULL, code = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{code}{Code dataset of old and new codes in a \code{data.table} format.
The dataset is the output after running \code{get_geo_recode()} function.}
}
\description{
Recode geographical codes to the current year.
Codes is based \code{\link[norgeo:track_change]{norgeo::track_change()}} function.
}
\examples{
\dontrun{
code <- get_geo_recode(con = geo$dbconn, type = "grunnkrets")
dt <- read_org("BEFOLKNING", aggregate = FALSE)
DT <- do_geo_recode(dt, code)
}
}
