% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file-each.R
\name{do_make_file_each}
\alias{do_make_file_each}
\title{Process Each File}
\usage{
do_make_file_each(spec, fgspec, aggregate, datacols, year, row, base)
}
\arguments{
\item{spec}{File specification}

\item{fgspec}{Filegroup specification}

\item{aggregate}{Logical value. Default is \code{TRUE}. Aggregate data according
to the specification in registration database. Global options with
\code{orgdata.aggregate}.}

\item{datacols}{Columnames to be kept}

\item{year}{Which year of georaphical code to use for recoding and
aggregating. If not specified, then current year will be used}

\item{row}{Select only specify row(s). Useful for debugging}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}
}
\description{
Process each file parallelly
}
