% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file-each.R
\name{do_make_file_each}
\alias{do_make_file_each}
\title{Process Each File}
\usage{
do_make_file_each(
  spec,
  fgspec,
  aggregate,
  datacols,
  year,
  row,
  base,
  duck = NULL
)
}
\arguments{
\item{spec}{File specification}

\item{fgspec}{Filegroup specification}

\item{aggregate}{Logical value. Default is \code{TRUE}. Aggregate data according
to the specification in registration database. Global options with
\code{orgdata.aggregate}.}

\item{datacols}{Columnames to be kept}

\item{year}{Which reference year to use for geograhical coding. If it is
missing then global option for \code{orgdata.year} will be used.}

\item{row}{Select specific row(s) numbers only. Useful for debugging. Please
read \href{https://helseprofil.github.io/orgdata/articles/debugging.html}{Debugging} article for detail.}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}

\item{duck}{R6 object for DuckDB}
}
\description{
Process each file according to the specification in Access database.
}
