% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{do_norgeo}
\alias{do_norgeo}
\title{Create Georgraphical Codes}
\usage{
do_norgeo(year = NULL, append = FALSE, write = FALSE, table = "tblGeo")
}
\arguments{
\item{year}{Year for the valid geographical codes}

\item{append}{Append the data to an existing table in the \code{orgdata.geo}}

\item{write}{Write table to the \code{orgdata.geo} database. It will overwrite
the table if it already exists}

\item{table}{Table name to be created in the database. Default is \code{tblGeo}}
}
\description{
Create a database for geographical codes to aggregate
data accordingly. Implementation of this function is base on \code{\link[norgeo:cast_geo]{norgeo::cast_geo()}}
function in \href{https://helseprofil.github.io/norgeo/}{norgeo} package.
}
