% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{do_recode}
\alias{do_recode}
\title{Recode Variables}
\usage{
do_recode(dt = NULL, spec = NULL, con = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Recode variables based on the specification in \code{tbl_Kode} ie. codebook.
\code{LESID} is the unique reference to recode variables.
}
