% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode-aggregate.R
\name{do_recode_aggregate}
\alias{do_recode_aggregate}
\title{Recode Category Representing All Groups}
\usage{
do_recode_aggregate(
  dt = NULL,
  spec = NULL,
  con = NULL,
  aggregate = getOption("orgdata.aggregate")
)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}

\item{aggregate}{Aggregate data according to the specification in registration database.
Default is \code{TRUE}. Global options with \code{orgdata.aggregate}.}
}
\description{
Recode the category for all after aggregating data based on the
specification in Codebook. For example when aggregating category type of diagnoses
\code{NA} will be produced to represent all type of diagnoses. This will basically
recode \code{NA} to a preferred value such as \code{ALL}.
}
\seealso{
Other recode functions: 
\code{\link{do_recode}()},
\code{\link{get_codebok_aggregate}()},
\code{\link{get_codebok}()}
}
\concept{recode functions}
