% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do-column-rename.R
\name{do_rename_col_standard}
\alias{do_rename_col_standard}
\title{Rename Standard Columns}
\usage{
do_rename_col_standard(file = NULL, df = NULL)
}
\arguments{
\item{file}{Input data}

\item{df}{Specifications data as a data.frame}
}
\description{
Renaming standard columns as in \code{getOption("orgdata.columns")}.
}
