% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{emoji}
\alias{emoji}
\title{Emoji}
\usage{
emoji(x = c("thumb", "smile", "sad"))
}
\arguments{
\item{x}{Emoji to choose ie. thumb, smile or sad}
}
\description{
Use emoji for fun :smile:
}
\examples{
emoji("smile")
}
