% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-group.R
\name{find_age_category}
\alias{find_age_category}
\title{Create age categories}
\usage{
find_age_category(dt = NULL, interval = NULL)
}
\arguments{
\item{dt}{Dataset}

\item{interval}{Age interval}
}
\description{
Create age categories either by age interval or specified age
categories. How to define the age categories in \code{EXTRA} column in Access is
shown in the example below.

Age categories can be specified as follows:
\itemize{
\item Specific interval eg. every 5 years. Interval with odd numbers will use minimum age of 0 and maximum age is 85+, while even number uses maximum age of 80+.
\item Specified interval eg. 0-18, 19-44, 45-64, 65-79, 80+
}
}
\examples{
\dontrun{
AgeCat(5) #Group age for every 5 years with min 0 and max 85+
AgeCat(10) #Group age for every 10 years with min 0 and max 80+
AgeCat(0, 19, 45, 65, 80) #Age group of 0-18, 19-44, 45-64, 65-79, 80+
}
}
\seealso{
Other extra arguments: 
\code{\link{do_extra_args_group}()},
\code{\link{do_extra_args}()},
\code{\link{get_extra_args_group}()},
\code{\link{get_extra_args}()}
}
\concept{extra arguments}
