% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multi-input.R
\name{find_column_multi}
\alias{find_column_multi}
\alias{find_column_multi_input}
\title{Columns with Multiple Inputs}
\usage{
find_column_multi(col = NULL, sep = c(",", ":", ";"))

find_column_multi_input(arg = NULL, y = NULL)
}
\arguments{
\item{col}{Column name in the database table}

\item{sep}{Symbols that seperate these arguments eg. "," or ":"}

\item{arg}{Name of argument}

\item{y}{Argument(s) as a vector}
}
\value{
A vector of the argument(s) that is seperated with \code{sep} argument

The value from the selected argument
}
\description{
Get all arguments in the columns that have multiple arguments with \code{\link[=find_column_multi]{find_column_multi()}}.
The output will be a vector that can be used in \code{\link[=find_column_multi_input]{find_column_multi_input()}} to get the value of the input.
}
\examples{
\dontrun{
args <- find_column_multi(spec$INNLESARG)
val <- find_column_multi_input("header", args)
}
}
