% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgeo.R
\name{geo_map_multi}
\alias{geo_map_multi}
\title{Granularity of Geographical Codes (multi-year)}
\usage{
geo_map_multi(from = NULL, to = NULL, write = FALSE, table = "tblGeo")
}
\arguments{
\item{from}{starting year}

\item{to}{ending year}

\item{write}{Write table to the \code{orgdata.geo} database. It will overwrite
the table if it already exists}

\item{table}{Table name to be created in the database. Default is \code{tblGeo}}
}
\description{
A wrapper around \verb{seeRD geo_map()}to generate a database granularity
of geographical codes to aggregate data accordingly, for multiple years.
}
