% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{get_aggregate}
\alias{get_aggregate}
\title{Get Aggregate Specification}
\usage{
get_aggregate(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification from \code{tbl_Filgruppe}}
}
\description{
Get the specification on how the data will be aggregated to
different geographical levels ie. county, manucipality, town etc.
}
