% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{get_codebok}
\alias{get_codebok}
\title{Codebook}
\usage{
get_codebok(spec = NULL, con = NULL)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{con}{Connection to database}
}
\description{
Get the codebook for recoding variables based on the
\code{FILGRUPPE} and \code{LESID} number. Specification group \code{ALLE} will be
used when neither \code{FILGRUPPE} nor \code{LESID} is specified.
}
\seealso{
Other recode functions: 
\code{\link{do_recode_aggregate}()},
\code{\link{do_recode_regexp}()},
\code{\link{do_recode}()},
\code{\link{get_codebok_aggregate}()},
\code{\link{get_codebok_regexp}()}
}
\concept{recode functions}
