% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-functions.R
\name{get_column_input}
\alias{get_column_input}
\title{Get Column Input}
\usage{
get_column_input(spec, col)
}
\arguments{
\item{spec}{Specifications data as a data.frame}

\item{col}{Column name in the database table}
}
\description{
Get the real column name in rawdata if it's not manually changed
as specified in the column \code{MANHEADER}. Only applicable for single input, else
check \code{\link[=find_column_multi]{find_column_multi()}} or \code{\link[=get_innlesarg]{get_innlesarg()}} function.
}
\examples{
\dontrun{
geo <- get_column_input(spec, "GEO")
}
}
