% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-multiple-args.R
\name{get_column_multi_args}
\alias{get_column_multi_args}
\alias{get_input_multi_args}
\title{Columns with Multiple Arguments}
\usage{
get_column_multi_args(col = NULL, sep = c(",", ":", ";"))

get_input_multi_args(arg = NULL, y = NULL)
}
\arguments{
\item{col}{Column name in the database table}

\item{sep}{Symbols that seperate these arguments eg. "," or ":"}

\item{arg}{Name of argument}

\item{y}{Argument(s) as a vector}
}
\description{
Get all arguments in the columns that have multiple arguments with \code{\link[=get_column_multi_args]{get_column_multi_args()}}.
The output will be a vector that can be used in \code{\link[=get_input_multi_args]{get_input_multi_args()}} to get the value of the input.
}
\examples{
\dontrun{
args <- get_column_multi_args(spec$INNLESARG)
val <- get_input_multi_args("header", args)
}
}
