% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesing.R
\name{get_column_name}
\alias{get_column_name}
\title{Get Column Name}
\usage{
get_column_name(df, col, type = c("character", "double", "integer"))
}
\arguments{
\item{df}{Data.frame extracted form registration database}

\item{col}{Column name in the database table}

\item{type}{Type of object as output ie. double, integer or character.
Default is character.}
}
\description{
Get the real column name in rawdata if it's not manually changed
as specified in the column \code{MANHEADER}. Only applicable for single input, else
check \code{\link[=find_column_multi]{find_column_multi()}} or \code{\link[=get_innlesarg]{get_innlesarg()}} function.
}
\examples{
\dontrun{
geo <- get_column_name(df, "GEO")
}
}
