% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard-column.R
\name{get_column_standard}
\alias{get_column_standard}
\title{Get Standard Columns}
\usage{
get_column_standard(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification of the standard columns in \code{tbl_Innlesing}}
}
\value{
A list with \code{old} and \code{new} columnnames
}
\description{
Standard columns names in rawdata to will be checked against
the standard names in options as in \code{getOptions("orgdata.columns")}.
}
