% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesarg.R
\name{get_innlesarg}
\alias{get_innlesarg}
\title{Get INNLESARG Specifications}
\usage{
get_innlesarg(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specifications data as a data.frame}
}
\value{
A list with the names and value of arguments
}
\description{
This function will access all possible arguments in columns \code{INNLESARG}
in table \emph{tbl_Innlesing}'. Symbol \code{"|"} \strong{MUST} be used to
separate arguments if there are more than one arguments. This symbol is right on top
of \code{TAB} button on the keyboard.

Arguments that can be used here depends on the file type. Please refer to \code{\link[=read_file]{read_file()}}.

For example:

\verb{header=TRUE | sheet=Ark1 | sep=,}
}
