% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innlesing.R
\name{get_innlesarg}
\alias{get_innlesarg}
\title{Get INNLESARG Specifications}
\usage{
get_innlesarg(arg, col, ...)
}
\arguments{
\item{arg}{Name of the arguments}

\item{col}{Column name in the table}

\item{...}{Other arguments eg. sep = ":"}
}
\value{
A character
}
\description{
This function will access all arguments in columns \code{INNLESARG} in table
\emph{tbl_Innlesing}' where there are possibility to
specify multiple arguments for example:
\verb{header=TRUE, ark=Sheet1}
}
\examples{
\dontrun{
input <- get_innlesarg("ark", spec$INNLESARG)
}
}
