% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-functions.R
\name{get_innlesarg}
\alias{get_innlesarg}
\title{Get INNLESARG Specifications}
\usage{
get_innlesarg(df = NULL)
}
\arguments{
\item{df}{Specifications data as a data.frame}
}
\value{
A list with the names and value of arguments
}
\description{
This function will access all possible arguments in columns \code{INNLESARG}
in table \emph{tbl_Innlesing}'. Comma \code{","} \strong{MUST} be used to
separate arguments if there are more than one argument.
For example:

\verb{header=TRUE, ark=Sheet1}
}
