% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-splitt.R
\name{get_split}
\alias{get_split}
\title{Get Split Column}
\usage{
get_split(group = NULL, con = NULL, spec = NULL)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{con}{Connection to database}

\item{spec}{Specification data from register database}
}
\value{
A list consist of to variable ie. \code{from} and \code{to}
indicating which column to split and what the new column name will be
}
\description{
Get the column to be splitted and the column name
for the new splitted columns.
If you already have the specification from \code{\link[=find_spec]{find_spec()}}, then the arguments
for \code{group} and \code{con} must be \code{NULL}.
}
