% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-column.R
\name{is_col_num_warn}
\alias{is_col_num_warn}
\title{Convert column to numeric with warning}
\usage{
is_col_num_warn(dt, cols, koblid = NULL)
}
\arguments{
\item{dt}{Dataset}

\item{cols}{Columns to be converted to numeric}

\item{koblid}{ID to connect file and spec}
}
\description{
Covert to numeric for columns that are expected to be numeric
and give warning and log when coercion
}
\keyword{internal}
