% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-db.R
\name{is_conn_db}
\alias{is_conn_db}
\title{Connect to Database}
\usage{
is_conn_db(dbname = NULL, db = c("kh", "geo", "raw"), .test = FALSE, ...)
}
\arguments{
\item{dbname}{Database filename with complete path}

\item{db}{Database file \code{kh} (Kommunehelse), \code{geo} (Geo code) or \code{raw} (Raw database)}

\item{.test}{Use for testing only}

\item{...}{Other arguments}
}
\description{
Alternative function to produce R6 object for connecting to the database
}
\keyword{internal}
