% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-process.R, R/process-org.R
\name{is_org_process}
\alias{is_org_process}
\title{Process Raw Data}
\usage{
is_org_process(.env = parent.frame(), verbose = getOption("orgdata.verbose"))

is_org_process(.env = parent.frame(), verbose = getOption("orgdata.verbose"))
}
\arguments{
\item{.env}{Enclosing environment}

\item{verbose}{Make processes explicit. Default is FALSE}
}
\value{
A dataset \code{.orgDT} with \code{data.table} format

A dataset \code{.orgDT} with \code{data.table} format
}
\description{
Implement the specifications to the raw data and deliver the output
to the enclosing environment.

Implement the specifications to the raw data and deliver the output
to the enclosing environment.
}
\keyword{internal}
