% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org-process.R, R/process-org.R
\name{is_org_process}
\alias{is_org_process}
\title{Process Raw Data}
\usage{
is_org_process(
  file,
  filespec,
  fgspec,
  con,
  verbose = getOption("orgdata.verbose")
)

is_org_process(
  file,
  filespec,
  fgspec,
  con,
  verbose = getOption("orgdata.verbose")
)
}
\arguments{
\item{file}{File rawdata}

\item{filespec}{Specification for a file from \code{tbl_Innlesing}}

\item{fgspec}{Specification for a file group}

\item{con}{Connection to the database}

\item{verbose}{Make processes explicit. Default is FALSE}
}
\value{
A dataset with \code{data.table} format

A dataset with \code{data.table} format
}
\description{
Implement the specifications to the raw data and deliver the output
to the enclosing environment.

Implement the specifications to the raw data and deliver the output
}
\keyword{internal}
