% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-db.R
\name{KHelse}
\alias{KHelse}
\title{Connecting to Database}
\description{
Connect to registration database to get all necessary information
on data source and cleaning specification. The driver is only applicable
to an Access Database. Calling the method \code{KHelse$new(YourFilePath)}
will create an object of R6 Class. Please refer to the
\href{https://helseprofil.github.io/orgdata/reference/KHelse.html#examples}{examples}.
}
\examples{

## ------------------------------------------------
## Method `KHelse$new`
## ------------------------------------------------

\dontrun{
kh <- KHelse$new(file.path(getOption("orgdata.drive"),
                           getOption("orgdata.folder.db"),
                           getOption("orgdata.db")))
kh$dbname
kh$db_close()
kh$db_connect()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{dbname}}{Database filename.}

\item{\code{dbtype}}{Database type of either Access or DuckDB}

\item{\code{dbyear}}{Production year. Only relevant for DuckDB}

\item{\code{dbconn}}{Database connection.}

\item{\code{tblname}}{Table name to be created in the database.}

\item{\code{tblvalue}}{Data to be inserted in the table \code{tblname}.
Data must be in a \code{data.frame} or \code{data.table} format.}

\item{\code{dbpath}}{Database path}

\item{\code{conn}}{Create database connection. Default is \code{TRUE}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KHelse-new}{\code{KHelse$new()}}
\item \href{#method-KHelse-db_connect}{\code{KHelse$db_connect()}}
\item \href{#method-KHelse-db_write}{\code{KHelse$db_write()}}
\item \href{#method-KHelse-db_read}{\code{KHelse$db_read()}}
\item \href{#method-KHelse-db_remove_table}{\code{KHelse$db_remove_table()}}
\item \href{#method-KHelse-db_close}{\code{KHelse$db_close()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-new"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-new}{}}}
\subsection{Method \code{new()}}{
Start connecting to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$new(
  dbname = NULL,
  dbtype = "Access",
  dbyear = NULL,
  dbpath = FALSE,
  conn = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbname}}{Database filename.}

\item{\code{dbtype}}{Database type eg. Access, SQLite, DuckDB etc.}

\item{\code{dbyear}}{Production year. This arg only relevant to raw database in DuckDB}

\item{\code{dbpath}}{Path to the database file}

\item{\code{conn}}{Create database connection. Default is \code{TRUE}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
kh <- KHelse$new(file.path(getOption("orgdata.drive"),
                           getOption("orgdata.folder.db"),
                           getOption("orgdata.db")))
kh$dbname
kh$db_close()
kh$db_connect()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-db_connect"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-db_connect}{}}}
\subsection{Method \code{db_connect()}}{
Reconnect to the database when \code{db_close} was used.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-db_write"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-db_write}{}}}
\subsection{Method \code{db_write()}}{
Write table to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_write(
  name = NULL,
  value = NULL,
  write = FALSE,
  append = FALSE,
  field.types = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name to be created in the database.}

\item{\code{value}}{The data to be inserted in the table.}

\item{\code{write}}{Write a table to the database. It will overwrite
the table if it already exists}

\item{\code{append}}{Append the data to an existing table in the database}

\item{\code{field.types}}{Type of data in specified column. Must be named as vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-db_read"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-db_read}{}}}
\subsection{Method \code{db_read()}}{
Read table and convert to data.table format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_read(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name to be created in the database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-db_remove_table"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-db_remove_table}{}}}
\subsection{Method \code{db_remove_table()}}{
Remove table in the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_remove_table(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name to be created in the database.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KHelse-db_close"></a>}}
\if{latex}{\out{\hypertarget{method-KHelse-db_close}{}}}
\subsection{Method \code{db_close()}}{
Close connection to the database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KHelse$db_close()}\if{html}{\out{</div>}}
}

}
}
