% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file.R
\name{make_file}
\alias{make_file}
\alias{lag_fil}
\title{Implement the Specifications}
\usage{
make_file(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = getOption("orgdata.year"),
  implicitnull = getOption("orgdata.implicit.null"),
  row = getOption("orgdata.debug.row"),
  base = getOption("orgdata.recode.base"),
  parallel = getOption("orgdata.parallel")
)

lag_fil(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = getOption("orgdata.year"),
  implicitnull = getOption("orgdata.implicit.null"),
  row = getOption("orgdata.debug.row"),
  base = getOption("orgdata.recode.base"),
  parallel = getOption("orgdata.parallel")
)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{aggregate}{Logical value. Default is \code{TRUE}. Aggregate data according
to the specification in registration database. Global options with
\code{orgdata.aggregate}.}

\item{save}{Save as \code{.csv} by activating \code{save_file()}. Default is \code{FALSE}}

\item{year}{Which year of georaphical code to use for recoding and
aggregating. If not specified, then current year will be used}

\item{implicitnull}{Logical value. Default is \code{TRUE} to add implicit null to
the dataset. Global options with \code{orgdata.implicit.null}.}

\item{row}{Select only specify row(s). Useful for debugging}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}

\item{parallel}{Logical or numeric value. With logical value \code{TRUE} it will
run with parallel using 50\% ie. 0.5 of local cores. User can decide other
percentage if needed. For example to use 75\% of the cores then specify as
\code{parallel = 0.75}. Nevertheless, maximum cores allowed is only 80\%.
Default value is \code{FALSE} ie. to use sequential processing}
}
\description{
Make a \code{csv} file with the specifications in the register
database and implement them to the raw data of the selected group of files
(\emph{filgruppe}). All files under the selected group will be affected
unless the \code{KOBLID} with argument \code{koblid} is specified.
Specifying \code{koblid} is useful especially for testing purposes.

The function \code{\link[=lag_fil]{lag_fil()}} is an alias to \code{\link[=make_file]{make_file()}}.
}
\examples{
\dontrun{
dt <- make_file("ENPERSON")
dt <- make_file("ENPERSON", koblid = 120:125, parallel = TRUE) #with parallel processing
}
}
\seealso{
Other filegroups functions: 
\code{\link{make_filegroups}()}
}
\concept{filegroups functions}
