% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file.R
\name{make_file}
\alias{make_file}
\alias{lag_fil}
\title{Implement the Specifications}
\usage{
make_file(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  implicitnull = getOption("orgdata.implicit.null"),
  row = getOption("orgdata.debug.row")
)

lag_fil(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  implicitnull = getOption("orgdata.implicit.null"),
  row = getOption("orgdata.debug.row")
)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{aggregate}{Aggregate data according to the specification in registration database.
Default is \code{TRUE}. Global options with \code{orgdata.aggregate}.}

\item{save}{Save as \code{.csv} by activating \code{save_file()}. Default is \code{FALSE}}

\item{year}{Which year of georaphical code to use for recoding and
aggregating. If not specified, then current year will be used}

\item{implicitnull}{Default is \code{TRUE} to add implicit null to the dataset. Global options
with \code{orgdata.implicit.null}.}

\item{row}{Select only specify row(s). Useful for debugging}
}
\description{
Make a \code{csv} file with the specifications in the register database and implement them
to the raw data of the selected group of files (\emph{filgruppe}). All files under
the selected group will be affected unless the \code{KOBLID} with
argument \code{koblid} is specified. Specifying \code{koblid} is useful
especially for testing purposes.

The function \code{\link[=lag_fil]{lag_fil()}} is an alias to \code{\link[=make_file]{make_file()}}.
}
\examples{
\dontrun{
make_file("BEFOLKNING")
make_file("BEFOLKNING", koblid = 19)
make_file("BEFOLKNING", koblid = c(15, 50))
}
}
\seealso{
Other filegroups functions: 
\code{\link{make_filegroups}()}
}
\concept{filegroups functions}
