% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file.R
\name{make_file}
\alias{make_file}
\alias{lag_fil}
\alias{mf}
\title{Implement the Specifications}
\usage{
make_file(
  group = NULL,
  koblid = NULL,
  aggregate = NULL,
  save = FALSE,
  year = NULL,
  implicitnull = NULL,
  row = NULL,
  base = NULL,
  parallel = deprecated(),
  raw = FALSE
)

lag_fil(
  group = NULL,
  koblid = NULL,
  aggregate = NULL,
  save = FALSE,
  year = NULL,
  implicitnull = NULL,
  row = NULL,
  base = NULL,
  parallel = deprecated(),
  raw = FALSE
)

mf(
  group = NULL,
  koblid = NULL,
  aggregate = NULL,
  save = FALSE,
  year = NULL,
  implicitnull = NULL,
  row = NULL,
  base = NULL,
  parallel = deprecated(),
  raw = FALSE
)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{aggregate}{Logical value. Default is \code{TRUE}. Aggregate data according
to the specification in registration database. Global options with
\code{orgdata.aggregate}.}

\item{save}{Logical value. Default is \code{FALSE}. To save as \code{.csv} format file
by activating \code{save_file()} function.}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then global option for \code{orgdata.year} will be used.}

\item{implicitnull}{Logical value. Default is \code{TRUE} to add implicit null to
the dataset. Global options with \code{orgdata.implicit.null}.}

\item{row}{Select specific row(s) numbers only. Useful for debugging. Please
read \code{Debugging} article for detail.}

\item{base}{Logical value. If \code{TRUE} then use year in the original data as the base
year to recode the geographical codes. Default is \code{FALSE} and use all
available codes in geo codebook}

\item{parallel}{Logical or numeric value. With logical value \code{TRUE} it will
run with parallel using 50\% ie. 0.5 of local cores. User can decide other
percentage if needed. For example to use 75\% of the cores then specify as
\code{parallel = 0.75}. Nevertheless, maximum cores allowed is only 80\%. Default
value is \code{FALSE} ie. to use sequential processing}

\item{raw}{Logical value. Default is \code{FALSE}. Read original raw data directly
from source file even if the dataset is already available in DuckDB without
the need to unmark \code{KONTROLLERT} in the Access database}
}
\description{
Make a \code{csv} file with the specifications in the Access register
database and implement them to the raw data of the selected group of files
ie. (\emph{filgruppe}). All files under the selected group will be affected
unless the \code{KOBLID} with argument \code{koblid} is specified.
Specifying \code{koblid} is useful especially for testing purposes.

The function \code{\link[=lag_fil]{lag_fil()}} is an alias to \code{\link[=make_file]{make_file()}}.
}
\examples{
\dontrun{
dt <- make_file("ENPERSON")
dt <- make_file("ENPERSON", raw = TRUE) #Skip DuckDB and read directly from original files
dt <- make_file("ENPERSON", koblid = 120:125, parallel = TRUE) #with parallel processing
}
}
\seealso{
Other filegroups functions: 
\code{\link{make_filegroups}()}
}
\concept{filegroups functions}
