% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgdata-package.R
\docType{package}
\name{orgdata-package}
\alias{orgdata}
\alias{orgdata-package}
\title{orgdata: Aggregating Original Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Cleaning, restructuring and aggregating original data
             usually with enumerical area codes to any geographical codes.
}
\details{
Aggregating Original Data (raw data) into different geographical
levels such as county, manucipality or town.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/helseprofil/orgdata}
  \item Report bugs at \url{https://github.com/helseprofil/orgdata/issues}
}

}
\author{
\strong{Maintainer}: Yusman Kamaleri \email{ybka@fhi.no} (\href{https://orcid.org/0000-0001-5014-3665}{ORCID})

}
\keyword{internal}
