% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{orgdata_logo}
\alias{orgdata_logo}
\title{Orgdata logo}
\usage{
orgdata_logo()
}
\value{
Orgdata logo
}
\description{
Display orgdata logo
}
\examples{
\dontrun{
if(interactive()){
 print.orgdata_logo()
 }
}
}
