% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-file.R
\name{read_file}
\alias{read_file}
\alias{lesfil}
\title{Read Data File}
\usage{
read_file(file = NULL, ...)

lesfil(file = NULL, ...)
}
\arguments{
\item{file}{Use FILID or a complete path of a filename}

\item{...}{All other arguments to be passed related to the file format}
}
\description{
Read rawdata either using \code{FILID} value or complete file path. It uses the \code{\link[=find_data]{find_data()}} generic method.
For a \code{ .csv } file, \code{\link[data.table:fread]{data.table::fread()}} is used and all other arguments
for \code{fread} function can be used. For a \code{ xlsx } or \code{ .xls } file
\code{\link[readxl:read_excel]{readxl::read_excel()}} function and all of its arguments.
}
\examples{
\dontrun{
# With FILID
DT <- read_file(file = 5)

# With filepath
rdata <- read_file(file = "/file/path/mydata.xlsx", sheet = "S3", range = cell_rows(1:4))
rdata <- read_file(file = "/file/path/mydata.csv", sep = ",", header = FALSE)
}
}
