% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{read_log}
\alias{read_log}
\title{Read Log File}
\usage{
read_log(name = NULL)
}
\arguments{
\item{name}{Log filename without file extention}
}
\description{
Read the log file from orgdata default path. To check the
default path use \code{orgdata:::is_orgdata_path()}.
}
\examples{
\dontrun{
read_log("code99")
}
}
