% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-org.R
\name{read_org}
\alias{read_org}
\alias{lesorg}
\title{Read and Run the Specifications}
\usage{
read_org(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  geo = NULL,
  val = NULL,
  ...
)

lesorg(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL,
  geo = NULL,
  val = NULL,
  ...
)
}
\arguments{
\item{group}{The group of files (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{aggregate}{Aggregate data according to the specification in registration database.
Default is FALSE. Use \code{options(orgdata.aggregate = TRUE)} to change globally.}

\item{save}{Save as \code{.csv} by activating \code{save_file()}. Default is \code{FALSE}}

\item{year}{Which year the georaphical code is valid for. If not specified, then
it will be base on the year in source data ie. column \code{AAR}}

\item{geo}{Which columname has geo codes. Only use in inactive mode.
Check status with \code{getOption("orgdata.active")}}

\item{val}{Which colunames has the value for quantity. Only use in an inactive mode.
Check status with \code{getOption("orgdata.active")}}

\item{...}{Additional parameters}
}
\description{
Read the specifications in the register database and implement them
to the selected group of files (\emph{filgruppe}). All files under
the selected group will be considered unless the \code{KOBLID} with
argument \code{koblid} is specified. Specifying \code{koblid} is useful for testing.

The function \code{\link[=lesorg]{lesorg()}} is an
alias to \code{\link[=read_org]{read_org()}}.
}
\examples{
\dontrun{
read_org("BEFOLKNING")
read_org("BEFOLKNING", koblid = 19)
read_org("BEFOLKNING", koblid = c(15, 50))
}
}
