% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-file.R
\name{read_raw}
\alias{read_raw}
\alias{lesraw}
\title{Implement the Specifications}
\usage{
read_raw(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL
)

lesraw(
  group = NULL,
  koblid = NULL,
  aggregate = getOption("orgdata.aggregate"),
  save = FALSE,
  year = NULL
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{read_raw} was renamed to \code{make_file()} to make it more clear
what the function is doing.
}
\keyword{internal}
