% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-file.R
\name{save_file}
\alias{save_file}
\alias{lagrefil}
\title{Save Data to CSV file}
\usage{
save_file(dt = NULL, group = NULL, path = NULL, date = FALSE, fgSpec = NULL)

lagrefil(dt = NULL, group = NULL, path = NULL, date = FALSE, fgSpec = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{group}{The group of files (\emph{filgruppe})}

\item{path}{Folder path to save the file. If not specified then the path
must to be specified in Access registration database}

\item{date}{Output file will be named with date and time}

\item{fgSpec}{File group specification from Access registration database}
}
\description{
Save data as a \code{.csv} format file with semicolon \verb{;} as seperation.
The file will be saved to the specified folder in \code{path} or as in Access regstration database
with the root as in \code{getOption("orgdata.folder.data")}. Use argument \code{save = TRUE}
in \code{read_raw()} will activate \code{save_file()} directly. Else you can call
\code{save_file()} to save the object output from \code{read_raw()}
}
\examples{
\dontrun{
 # Save file directly
 read_raw("BEFOLKNING", save = TRUE)

 # Two steps
 DF <- read_raw("BEFOLKNING")
 save_file(DF, "BEFOLKNING")
}
}
