% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-file.R
\name{save_file}
\alias{save_file}
\alias{lagfil}
\title{Save Data to CSV file}
\usage{
save_file(dt = NULL, group = NULL, date = FALSE)

lagfil(dt = NULL, group = NULL, date = FALSE)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{group}{The group of files (\emph{filgruppe})}

\item{date}{Output file will be named with date and time}
}
\description{
Save data as a \code{.csv} format file with semicolon \verb{;} as seperation.
The file will be saved to the specified folder
as in \code{getOption("orgdata.folder.output")}. Use argument \code{save = TRUE}
in \code{read_raw()} will activate \code{save_file()} directly. Else you can call
\code{save_file()} to save the object output from \code{read_raw()}
}
\examples{
\dontrun{
 opitions(orgdata.aggregate = TRUE)
 DF <- read_raw("BEFOLKNING", save = TRUE)

 # Two steps
 DF <- read_raw("BEFOLKNING")
 save_file(DF, "BEFOLKNING")
}
}
