% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-file.R
\name{save_file}
\alias{save_file}
\alias{lagre_fil}
\title{Save Data to CSV file}
\usage{
save_file(
  dt = NULL,
  name = NULL,
  path = NULL,
  date = FALSE,
  fgSpec = NULL,
  sep = ";",
  parquet = FALSE,
  ...
)

lagre_fil(
  dt = NULL,
  name = NULL,
  path = NULL,
  date = FALSE,
  fgSpec = NULL,
  sep = ";",
  parquet = FALSE,
  ...
)
}
\arguments{
\item{dt}{data}

\item{name}{Filename for the \code{.csv} file or filegroup name}

\item{path}{Folder path to save the file. If \code{name} is a valide filegroup
\emph{(FILGRUPPE)} then use the specified \code{UTMAPPE} in Access registration
database else file will be saved in default folder
\verb{C:\\Users\\YourUserName\\orgdata_logs}. The default folder will be created if
not exist.}

\item{date}{Logical value. If TRUE then use date and time as part of the filename. Default is FALSE.}

\item{fgSpec}{File group specification from Access registration database}

\item{sep}{The separator between columns. Default is \code{";"}}

\item{parquet}{Logical value, to additionally save a \code{.parquet} file}

\item{...}{Other arguments for \code{data.table::fwrite}}

\item{csvname}{name of csv-file, to be used as basis for the .qdata file name}
}
\description{
Save data as a \code{.csv} format file with semicolon \verb{;} as
seperator. The file will be saved to the specified folder in \code{path} or as
in Access regstration database with the root as in
\code{getOption("orgdata.folder.data")}. Use argument \code{save = TRUE} in
\code{make_file()} will activate \code{save_file()} directly. Else you can call
\code{save_file()} to save the object output from \code{make_file()}. This function
is a wrapper to \code{data.table:fwrite()}. Additionally, a \code{.qs}-file is saved,
which will be the main format in the future.

Saves a .qdata file in addition to the .csv-file
All columns are first converted to character, which is needed in the next steps
}
\examples{
\dontrun{
 # Save file directly
 make_file("BEFOLKNING", save = TRUE)

 # Two steps
 DF <- make_file("BEFOLKNING")
 save_file(DF, "BEFOLKNING")

 # Save with different name and specified folder
 save_file(DF, name = "myCSVfile", path = "C:/MyFolder")
}
}
