% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-file.R
\name{save_file}
\alias{save_file}
\alias{lagre_fil}
\title{Save Data to CSV file}
\usage{
save_file(dt = NULL, name = NULL, path = NULL, date = FALSE, fgSpec = NULL)

lagre_fil(dt = NULL, name = NULL, path = NULL, date = FALSE, fgSpec = NULL)
}
\arguments{
\item{dt}{Input data preferrably in a \code{data.table} format but is not mandatory}

\item{name}{Filename for the \code{.csv} file or filegroup name}

\item{path}{Folder path to save the file. If \code{name} is
a valide filegroup \emph{(FILGRUPPE)} then use the specified \code{UTMAPPE}
in Access registration database else file will be saved in default folder
\verb{C:\\Users\\YourUserName\\orgdata_files}. The default folder will be
created if not exist.}

\item{date}{Use date and time as part of the filename}

\item{fgSpec}{File group specification from Access registration database}
}
\description{
Save data as a \code{.csv} format file with semicolon \verb{;} as
seperation. The file will be saved to the specified folder in \code{path} or as
in Access regstration database with the root as in
\code{getOption("orgdata.folder.data")}. Use argument \code{save = TRUE} in
\code{make_file()} will activate \code{save_file()} directly. Else you can call
\code{save_file()} to save the object output from \code{make_file()}
}
\examples{
\dontrun{
 # Save file directly
 make_file("BEFOLKNING", save = TRUE)

 # Two steps
 DF <- make_file("BEFOLKNING")
 save_file(DF, "BEFOLKNING")

 # Save with different name and specified folder
 save_file(DF, name = "myCSVfile", path = "C:/MyFolder")
}
}
