% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-data.R
\name{see_data}
\alias{see_data}
\title{See Structured Data in the Data Warehouse}
\usage{
see_data(
  group = NULL,
  koblid = NULL,
  year = NULL,
  action = c("read", "delete")
)
}
\arguments{
\item{group}{The filegroup name (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then global option for \code{orgdata.year} will be used.}

\item{action}{To read or delete the data in the warehouse. Default is \code{read}.}
}
\description{
See the original data that have been clean and structured i the
data warehouse. Data is saved in the warehouse when the column
\emph{KONTROLLERT} is marked in the original file table in Access. This means
the dataset has been cleaned and recoded as specified in \emph{INNLESING} table
in Access registration database.
}
\examples{
\dontrun{
dt <- see_data("LESEFERD", koblid = 134)
}
}
