% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-data.R
\name{see_data}
\alias{see_data}
\title{Access Structured Data in Data Warehouse}
\usage{
see_data(
  group = NULL,
  koblid = NULL,
  year = NULL,
  action = c("read", "delete")
)
}
\arguments{
\item{group}{The filegroup name (\emph{filgruppe})}

\item{koblid}{Specify one or multiple KOBLID. Use \code{"all"} to delete all data in warehouse for seleted filegroup.}

\item{action}{To read or delete the data in the warehouse. Default is \code{read}.}
}
\description{
See or delete the original data in data warehouse that have been cleaned and structured.
Data is saved in the warehouse when column
\emph{KONTROLLERT} is marked in the original file table in Access database. This indicates
that dataset has been checked and recoded as specified in \emph{INNLESING} table
in Access registration database.
}
\examples{
\dontrun{
dt <- see_data("LESEFERD", koblid = 134)
dt <- see_data("ENPERSON", koblid = "all", action = "delete")
dt <- see_data("ENPERSON", koblid = 267:270, action = "delete")
}
}
