% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-file.R
\name{see_file}
\alias{see_file}
\alias{se_fil}
\title{See Column Categories}
\usage{
see_file(dt = NULL, ...)

se_fil(dt = NULL, ...)
}
\arguments{
\item{dt}{Dataset of type \code{data.frame} or \code{data.table}}

\item{...}{Columnames or column index to be displayed}
}
\description{
See what categories available in the dataset.
}
\examples{
\dontrun{
# Use columnames
see_file(DT, KJONN, UTDANN, LANDSSB)

# Use column index
see_file(DT, c(2,5))
see_file(DT, c(1:3))
see_file(DT, c(2, 4, 7:8))
}
}
