% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-org.R
\name{see_org}
\alias{see_org}
\title{See Original Data in the Database}
\usage{
see_org(group = NULL, koblid = NULL, year = NULL, action = c("read", "delete"))
}
\arguments{
\item{group}{The filegroup name (\emph{filgruppe})}

\item{koblid}{\code{KOBLID} from table \emph{tbl_Koble}}

\item{year}{Which year the geograhical codes to be recoded to. If it is empty
then global option for \code{orgdata.year} will be used.}

\item{action}{To read or delete the data in the database. Default is \code{read}.}
}
\description{
See the original data that are saved in the org database when
the column \emph{KONTROLLERT} is marked. This means the dataset has been cleaned
and recoded as specified in \emph{INNLESING} table in Access registration
database.
}
\examples{
\dontrun{
dt <- see_org("LESEFERD", koblid = 134)
}
}
