% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{update_globs}
\alias{update_globs}
\alias{up_opt}
\title{Update Global Options}
\usage{
update_globs(package = c("orgdata", "orgcube", "khfunctions", "qualcontrol"))

up_opt(package = c("orgdata", "orgcube", "khfunctions", "qualcontrol"))
}
\arguments{
\item{package}{"orgdata" (default), "orgcube", "khfunctions", or "qualcontrol"}
}
\description{
Update global options based on the file in config repo
}
