% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{update_orgdata}
\alias{update_orgdata}
\alias{upgrade_orgdata}
\title{Update package}
\usage{
update_orgdata(...)

upgrade_orgdata(...)
}
\arguments{
\item{...}{Other arguments to \code{remotes::install_github()}}
}
\description{
Update orgdata directly with Github repo. Default is \code{main} branch.
}
\examples{
\dontrun{
update_orgdata() #use default
update_orgdata(ref = "dev") #to upgrade with dev branch
}
}
